Separate logic from component
